+++ R/as_survey_design.r
@@ -12,6 +12,11 @@
 #' If provided a \code{survey.design2} object from the survey package,
 #' it will turn it into a srvyr object, so that srvyr functions will work with it
 #'
+#' There is also limited and experimental support for databases using dplyr's \code{tbl_sql}
+#' objects. Not all operations are available for these objects. See
+#' \code{vignette("databases", package = "dplyr")} for more information on setting
+#' up databases in dplyr.
+#'
 #' @export
 #' @param .data A data frame (which contains the variables specified below)
 #' @param ids Variables specifying cluster ids from largest level to smallest level
@@ -30,6 +35,8 @@
 #' approximation. An object of class ppsmat to use the Horvitz-Thompson estimator.
 #' @param variance For pps without replacement, use variance="YG" for the Yates-Grundy estimator
 #' instead of the Horvitz-Thompson estimator
+#' @param uid Required for databases only, variables that uniquely identify the
+#' observations of your survey.
 #' @param ... ignored
 #' @return An object of class \code{tbl_svy}
 #' @examples
@@ -84,8 +91,8 @@ as_survey_design <- function(.data, ...) {
 as_survey_design.data.frame <-
   function(.data, ids = NULL, probs = NULL, strata = NULL,
            variables = NULL, fpc = NULL, nest = FALSE,
-           check_strata = !nest,weights = NULL, pps = FALSE,
-           variance = c("HT", "YG"), ...) {
+           check_strata = !nest, weights = NULL, pps = FALSE,
+           variance = c("HT", "YG"), uid = NULL, ...) {
 
   if (!missing(ids)) {
     ids <- lazy_parent(ids)
@@ -96,13 +103,17 @@ as_survey_design.data.frame <-
   if (!missing(fpc)) fpc <- helper(lazy_parent(fpc), .data)
   if (!missing(weights)) weights <- helper(lazy_parent(weights), .data)
   if (!missing(variables)) variables <- helper(lazy_parent(variables), .data)
+  if (!missing(uid)) uid <- helper(lazy_parent(uid), .data)
 
   as_survey_design_(.data, ids, probs = probs, strata = strata,
                     variables = variables, fpc = fpc, nest = nest,
                     check_strata = check_strata, weights = weights, pps = pps,
-                    variance = variance)
+                    variance = variance, uid = uid)
 }
 
+#' @export
+#' @rdname as_survey_design
+as_survey_design.tbl_sql <- as_survey_design.data.frame
 
 #' @export
 #' @rdname as_survey_design
@@ -116,7 +127,7 @@ as_survey_design.survey.design2 <- function(.data, ...) {
 as_survey_design_ <- function(.data, ids = NULL, probs = NULL, strata = NULL,
                               variables = NULL, fpc = NULL, nest = FALSE,
                               check_strata = !nest,weights = NULL, pps = FALSE,
-                              variance = c("HT", "YG")) {
+                              variance = c("HT", "YG"), uid = NULL) {
 
 
   # svydesign expects ~0 instead of NULL if no ids are included
@@ -126,22 +137,29 @@ as_survey_design_ <- function(.data, ids = NULL, probs = NULL, strata = NULL,
   } else {
     ids_call <- dplyr::select_(.data, .dots = ids)
   }
-  # Need to convert to data.frame to appease survey package and also not
-  # send NULL to dplyr::select
-  survey_selector <- function(x) {
-    if (!is.null(x)) data.frame(dplyr::select_(.data, .dots = x)) else NULL
+
+  # Databases require uid
+  if (inherits(.data, "tbl_lazy")) {
+    # Databases require uid
+    if (missing(uid) || is.null(uid)) {
+      stop("Database backed surveys require a uid.")
+    } else {
+      uid_names <- get_uid_names(length(uid))
+      .data <- uid_rename(.data, uid, uid_names)
+      attr(.data, "order_var") <- uid_names
+    }
   }
 
   out <- survey::svydesign(data = .data,
                            ids = ids_call,
-                           probs = survey_selector(probs),
-                           strata = survey_selector(strata),
-                           variables = survey_selector(variables),
-                           fpc = survey_selector(fpc),
-                           weights = survey_selector(weights),
+                           probs = survey_selector(.data, probs),
+                           strata = survey_selector(.data, strata),
+                           variables = survey_selector(.data, variables),
+                           fpc = survey_selector(.data, fpc),
+                           weights = survey_selector(.data, weights),
                            nest = nest, check.strata = check_strata, pps = pps,
                            variance = variance)
 
   as_tbl_svy(out, list(ids = ids, probs = probs, strata = strata, fpc = fpc,
-                       weights = weights))
+                       weights = weights), uid = survey_selector(.data, uid))
 }