+++ R/manip.r
@@ -13,26 +13,20 @@ mutate_.tbl_svy <- function(.data, ..., .dots) {
 #' @export
 select_.tbl_svy <- function(.data, ..., .dots) {
   dots <- lazyeval::all_dots(.dots, ...)
-  vars <- dplyr::select_vars_(names(.data$variables), dots,
-                       include = attr(.data, "group_vars"))
+  vars <- dplyr::select_vars_(dplyr::tbl_vars(.data$variables), dots,
+                       include = c(attr(.data, "group_vars"),
+                                   attr(.data$variables, "order_var")))
 
   .data$variables <- select_(.data$variables, .dots = vars)
 
-  # Also rename survey_vars, group_vars, and the structures in the
-  # svydesign2 object
-  .data <- rename_special_vars(.data, vars)
   .data
 }
 
 #' @export
 rename_.tbl_svy <- function(.data, ..., .dots) {
   dots <- lazyeval::all_dots(.dots, ...)
-  vars <- dplyr::rename_vars_(names(.data$variables), dots)
+  vars <- dplyr::rename_vars_(dplyr::tbl_vars(.data$variables), dots)
   .data$variables <- rename_(.data$variables, .dots = vars)
-
-  # Also rename survey_vars, group_vars, and the structures in the
-  # svydesign2 object
-  .data <- rename_special_vars(.data, vars)
 
   .data
 }
@@ -40,49 +34,7 @@ rename_.tbl_svy <- function(.data, ..., .dots) {
 #' @export
 filter_.tbl_svy <- function(.data, ..., .dots) {
   dots <- lazyeval::all_dots(.dots, ...)
-
-  # There's probably a better way to do this... But I need to use
-  # survey::subset because I want to make sure that I recalculate the
-  # survey_vars correctly. Create a variable with the row numbers, run dplyr
-  # on the variables data.frame and then pass the row_numbers that are kept
-  # into survey::svydesign2 `[`
-  row_numbers <- .data$variables
-  total_num_rows <- nrow(row_numbers)
-  row_numbers <-
-    dplyr::mutate(row_numbers, `___row_nums` = seq(1, total_num_rows))
-  row_numbers <- dplyr::filter_(row_numbers, .dots = dots)
-  row_numbers <- row_numbers$`___row_nums`
-
-  .data[row_numbers, ]
-}
-
-# Helper to rename variables stored in survey_vars and svydesign2 stucture
-rename_special_vars <- function(svy, var_list) {
-  renamed_vars <- var_list[var_list != names(var_list)]
-  svars <- as.character(survey_vars(svy))
-
-  for (iii in seq_along(renamed_vars)) {
-    this_var <- renamed_vars[iii]
-
-    # Make changes in the survey_vars structure
-    svars <- lapply(svars, function(x) {
-      x[x == this_var] <- names(this_var)
-      x
-    })
-
-    # Make changes in actual sydesign2 object's structures
-    names(svy$cluster)[names(svy$cluster) == this_var] <- names(this_var)
-    if(svy$has.strata) {
-      names(svy$strata)[names(svy$strata) == this_var] <- names(this_var)
-    }
-    names(svy$allprob)[names(svy$allprob) == this_var] <- names(this_var)
-    attr(svy$fpc$popsize, "dimnames")[[2]][
-      attr(svy$fpc$popsize, "dimnames")[[2]] == this_var
-      ] <- names(this_var)
-  }
-
-  survey_vars(svy) <- svars
-  svy
+  subset_svy_vars(.data, dots)
 }