+++ R/utils.r
@@ -50,16 +50,16 @@ helper_list <- function(x, .data) {
   if(x[[1]] != "list" || length(x) > 3) {
     stop("as_survey_twophase requies a list of 2 sets of variables")
   }
-  name1 <- unname(dplyr::select_vars_(names(.data), x[[2]]))
+  name1 <- unname(dplyr::select_vars_(dplyr::tbl_vars(.data), x[[2]]))
   name1 <- if (length(name1) == 0) NULL else name1
-  name2 <- unname(dplyr::select_vars_(names(.data), x[[3]]))
+  name2 <- unname(dplyr::select_vars_(dplyr::tbl_vars(.data), x[[3]]))
   name2 <- if (length(name2) == 0) NULL else name2
   list(name1, name2)
 }
 
 # Need to turn bare variable to variable names (when not in list)
 helper <- function(x, .data) {
-  unname(dplyr::select_vars_(names(.data), x))
+  unname(dplyr::select_vars_(dplyr::tbl_vars(.data), x))
 }
 
 # From http://stackoverflow.com/questions/7963898
@@ -80,3 +80,41 @@ substr_right <- function(x, n){
 #' @importFrom magrittr %>%
 #' @usage lhs \%>\% rhs
 NULL
+
+# Need to convert to data.frame to appease survey package and also not
+# send NULL to dplyr::select
+survey_selector <- function(.data, x) {
+  if (!is.null(x)) {
+
+    if (inherits(.data, "tbl_lazy")) {
+      out <- dplyr::select_(.data, .dots = c(attr(.data, "order_var"), x))
+      out <- ordered_collect(out)
+    } else {
+      out <- dplyr::select_(.data, .dots = x)
+    }
+    out <- data.frame(out)
+  } else {
+    out <- NULL
+  }
+  out
+}
+
+# survey::twophase doesn't work with values, needs to be formula of
+# variable names
+# Change list of variable names to formulas
+list_to_formula <- function(x) {
+  if (!is.null(x)) {
+    lapply(x, function(y) nullable(survey::make.formula, y))
+  } else NULL
+}
+
+# From survey:::is.calibrated()
+is.calibrated <- function(design)
+{
+  !is.null(design$postStrata)
+}
+
+# From survey:::is.pps()
+is.pps <- function(x) {
+  if (is.null(x$pps)) FALSE else (x$pps != FALSE)
+}